+++ CHANGELOG.md
@@ -5,6 +5,10 @@ documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Fixed
+- Integration test polluting `$HOME` for docker client stub
+
 ## [0.0.47] - 2020-07-23
 ### Added
 - version info to `--debug` above the stacktrace/s (thanks [Andreas Sundqvist])